chore(Figure): 영역확인 메서드 리턴 타입 변경
다형성처리에 맞게끔 타입 변경
int -> double

+++ src/main/java/coordinate/domain/Figure.java
@@ -2,7 +2,8 @@ package coordinate.domain;
 
 
 public interface Figure {
-    int getArea();
+
+    double getArea();
 
     String getAreaString();
 }
